% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothCutsur.R
\name{ToothCutsur}
\alias{ToothCutsur}
\title{ToothCutsur}
\usage{
ToothCutsur(outline, scale = 1.2)
}
\arguments{
\item{outline}{3D matrix: cervical outline}

\item{scale}{numeric: magnification factor}
}
\value{
mesh cutting surface
}
\description{
Create a 3D mesh from the coordinates of the cervical outline and its barycenter
}
\author{
Antonio Profico
}
