% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothExData.R
\docType{data}
\name{ToothExData}
\alias{ToothExData}
\title{example dataset}
\usage{
data(ToothExData)
}
\description{
Neanderthal example dataset (upper left canine): enamel, dentin and pulp meshes plus a curve along the cervical outline and four landmarks on the lingual, mesial, buccal and distal margins
}
\author{
Antonio Profico
}
\keyword{Toothnroll}
