% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothExport.R
\name{ToothExport}
\alias{ToothExport}
\title{ToothExport}
\usage{
ToothExport(input, id, file)
}
\arguments{
\item{input}{list: output from ToothAlignement}

\item{id}{character: label name}

\item{file}{character: name the output file}
}
\description{
Export the output from ToothAlignement
}
\author{
Antonio Profico
}
