% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothImport.R
\name{ToothImport}
\alias{ToothImport}
\title{ToothImport}
\usage{
ToothImport(file)
}
\arguments{
\item{file}{character: name of input file}
}
\value{
Mesh1: mesh of the aligned mesh1

Mesh2: mesh of the aligned mesh2

Mesh3: mesh of the aligned mesh3

B.Length: biomechanical lenght

Landmarks: landmark coordinates

Outline: outline coordinates

Margins: position of the margins along the outline
}
\description{
Import the output from ToothAlignement
}
\author{
Antonio Profico
}
