% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothMeshIntersect.R
\name{ToothMeshIntersect}
\alias{ToothMeshIntersect}
\title{ToothMeshIntersect}
\usage{
ToothMeshIntersect(refmesh, tarmesh)
}
\arguments{
\item{refmesh}{object of class mesh3d: reference mesh (cutting surface)}

\item{tarmesh}{object of class mesh3d: target mesh (surface to be cut)}
}
\value{
meshUp upper region of the cut mesh

meshLo lower region of the cut mesh
}
\description{
Remove intersection between two aligned meshes
}
\examples{
data("ToothExData")
require(rgl)
require(Rvcg)
Enamel<-ToothExData$enamel
Dentin<-ToothExData$dentin
Pulp<-ToothExData$pulp
outline<-ToothExData$outline
set<-ToothExData$set
AlignMeshes<-ToothAlignment(mesh1=Enamel,mesh2=Dentin,mesh3=Pulp,set,outline)
cutSurface<-ToothCutsur(AlignMeshes$outline,scale=1.2)
cutSurface2<-vcgClean(vcgUniformRemesh(vcgClean(cutSurface),voxelSize=0.1,multiSample = TRUE))
InterMeshes<-ToothMeshIntersect(cutSurface2,AlignMeshes$almesh2$mesh)
if(interactive()){
 wire3d(cutSurface,col=1)
 wire3d(InterMeshes$meshUp,col=2)
 wire3d(InterMeshes$meshLo,col=3)
}
}
\author{
Antonio Profico
}
