% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toothnroll.R
\docType{package}
\name{Toothnroll-package}
\alias{Toothnroll-package}
\alias{Toothnroll}
\title{Equiangular semilandmarks, virtual unrolling, 2D and 3D maps of enamel and dentine thickess}
\description{
Tool to process dental meshes (shape data, virtual unrolling, and morphometric maps)
}
\author{
Antonio Profico
}
