% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothRegradius.R
\name{ToothRegradius}
\alias{ToothRegradius}
\title{ToothRegradius}
\usage{
ToothRegradius(mat, center, n, direction = c("c", "a"))
}
\arguments{
\item{mat}{a kx2 matrix}

\item{center}{coordinates of the center from which the calculation of regular radius started}

\item{n}{number of points}

\item{direction}{specify direction of equiangular landmarks: "c"=clockwise; "a"=anticlockwise}
}
\value{
V2 position of landmarks equiangular spaced
}
\description{
Wrapper of the function regularradius written by Julien Claude (Morphometrics with R)
}
\author{
Julien Claude, Antonio Profico
}
