% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothThickness.R
\name{ToothThickness}
\alias{ToothThickness}
\title{ToothThickness}
\usage{
ToothThickness(tooth.shape)
}
\arguments{
\item{tooth.shape}{list: tooth.shape object}
}
\value{
sect_thickness_eq dental thickness (equiangular method)

sect_thickness_cp dental thickness (closest point method)

sect_Rthickness_eq relative dental thickness (equiangular method)

sect_Rthickness_cp relative dental thickness (closest point method)

ALPM_thickness dental thickness at ALPM quadrants

tooth.shape tooth.shape object
}
\description{
Tool for the extraction of equiangular landmarks on the selected dental mesh
}
\author{
Antonio Profico; Mathilde Augoyard
}
