% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToothVariations.R
\name{ToothVariations}
\alias{ToothVariations}
\title{ToothVariations}
\usage{
ToothVariations(
  PCA,
  scores,
  PC,
  asp = 2,
  pal = blue2green2red(101),
  meanmap = TRUE
)
}
\arguments{
\item{PCA}{list: output from ToothPCA}

\item{scores}{numeric: principal component value}

\item{PC}{numeric: principal component eigenvalue}

\item{asp}{numeric: x,y ratio of the morphometric map}

\item{pal}{character vector: vector of colors}

\item{meanmap}{logical: if TRUE the mean map corresponds to the real mean morphometric maps, otherwise the mean map is 0}
}
\value{
XYZ data.frame of morphometric map variation
}
\description{
Calculate and return morphometric map variation
}
\author{
Antonio Profico; Mathilde Augoyard
}
