% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCcrown.R
\docType{data}
\name{UCcrown}
\alias{UCcrown}
\title{example dataset}
\usage{
data(UCcrown)
}
\description{
list of ToothShape (modern human and neanderthal upper canines)
}
\author{
Antonio Profico, Mathilde Augoyard
}
\keyword{Toothnroll}
