% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCroot.R
\docType{data}
\name{UCroot}
\alias{UCroot}
\title{example dataset}
\usage{
data(UCroot)
}
\description{
list of ToothShape (modern human and neanderthal upper canines)
}
\author{
Antonio Profico, Mathilde Augoyard
}
\keyword{Toothnroll}
