% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI1crown.R
\docType{data}
\name{UI1crown}
\alias{UI1crown}
\title{example dataset}
\usage{
data(UI1crown)
}
\description{
list of ToothShape (modern human and neanderthal upper central incisors)
}
\author{
Antonio Profico, Mathilde Augoyard
}
\keyword{Toothnroll}
