% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI1root.R
\docType{data}
\name{UI1root}
\alias{UI1root}
\title{example dataset}
\usage{
data(UI1root)
}
\description{
list of ToothShape (modern human and neanderthal upper central incisors)
}
\author{
Antonio Profico, Mathilde Augoyard
}
\keyword{Toothnroll}
