% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI2crown.R
\docType{data}
\name{UI2crown}
\alias{UI2crown}
\title{example dataset}
\usage{
data(UI2crown)
}
\description{
list of ToothShape (modern human and neanderthal upper lateral incisors)
}
\author{
Antonio Profico, Mathilde Augoyard
}
\keyword{Toothnroll}
