% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI2root.R
\docType{data}
\name{UI2root}
\alias{UI2root}
\title{example dataset}
\usage{
data(UI2root)
}
\description{
list of ToothShape (modern human and neanderthal upper lateral incisors)
}
\author{
Antonio Profico, Mathilde Augoyard
}
\keyword{Toothnroll}
