% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URI1_tooth.R
\docType{data}
\name{URI1_tooth}
\alias{URI1_tooth}
\title{example dataset}
\usage{
data(URI1_tooth)
}
\description{
3D dental meshes of a modern human upper right central incisor
}
\author{
Antonio Profico, Mathilde Augoyard
}
\keyword{Toothnroll}
